feat: error codes diagnosis support
