+++ go.mod
@@ -2,6 +2,8 @@ module github.com/speakeasy-api/speakeasy
 
 go 1.23.0
 
+replace github.com/speakeasy-api/speakeasy-core => ../speakeasy-core
+
 // Can be removed once https://github.com/pb33f/libopenapi/pull/319 is merged and the dependency is updated here
 replace github.com/pb33f/libopenapi => github.com/speakeasy-api/libopenapi v0.0.0-20240814113924-cc96d2bc2826
 
@@ -31,7 +33,7 @@ require (
 	github.com/speakeasy-api/openapi-overlay v0.9.0
 	github.com/speakeasy-api/sdk-gen-config v1.23.2
 	github.com/speakeasy-api/speakeasy-client-sdk-go/v3 v3.13.1
-	github.com/speakeasy-api/speakeasy-core v0.15.5
+	github.com/speakeasy-api/speakeasy-core v0.15.6
 	github.com/speakeasy-api/speakeasy-proxy v0.0.2
 	github.com/spf13/cobra v1.8.1
 	github.com/spf13/pflag v1.0.5
