+++ internal/suggest/errorCodes/testData/nameConflict.yaml
