+++ internal/suggest/errorCodes/testData/nameConflict_expected.yaml
