+++ internal/suggest/errorCodes/testData/simple.yaml
